Move some namespaces
