+++ src/Bundle/test/src/Grid/AuthorWithBooksWithFetchJoinCollectionEnabled.php
@@ -14,9 +14,9 @@ declare(strict_types=1);
 namespace App\Grid;
 
 use App\Entity\Author;
-use Sylius\Bundle\GridBundle\AbstractGrid;
 use Sylius\Bundle\GridBundle\Builder\GridBuilderInterface;
-use Sylius\Bundle\GridBundle\ResourceAwareGridInterface;
+use Sylius\Bundle\GridBundle\Grid\AbstractGrid;
+use Sylius\Bundle\GridBundle\Grid\ResourceAwareGridInterface;
 
 final class AuthorWithBooksWithFetchJoinCollectionEnabled extends AbstractGrid implements ResourceAwareGridInterface
 {
