+++ src/Bundle/test/src/Grid/AuthorWithBooksWithUseOutputWalkersEnabled.php
@@ -14,10 +14,10 @@ declare(strict_types=1);
 namespace App\Grid;
 
 use App\Entity\Author;
-use Sylius\Bundle\GridBundle\AbstractGrid;
 use Sylius\Bundle\GridBundle\Builder\Field\StringField;
 use Sylius\Bundle\GridBundle\Builder\GridBuilderInterface;
-use Sylius\Bundle\GridBundle\ResourceAwareGridInterface;
+use Sylius\Bundle\GridBundle\Grid\AbstractGrid;
+use Sylius\Bundle\GridBundle\Grid\ResourceAwareGridInterface;
 
 final class AuthorWithBooksWithUseOutputWalkersEnabled extends AbstractGrid implements ResourceAwareGridInterface
 {
