+++ src/Bundle/test/src/Grid/BookGrid.php
@@ -16,11 +16,11 @@ namespace App\Grid;
 use App\Entity\Author;
 use App\Entity\Book;
 use App\Entity\Nationality;
-use Sylius\Bundle\GridBundle\AbstractGrid;
 use Sylius\Bundle\GridBundle\Builder\Field\StringField;
 use Sylius\Bundle\GridBundle\Builder\Filter\Filter;
 use Sylius\Bundle\GridBundle\Builder\GridBuilderInterface;
-use Sylius\Bundle\GridBundle\ResourceAwareGridInterface;
+use Sylius\Bundle\GridBundle\Grid\AbstractGrid;
+use Sylius\Bundle\GridBundle\Grid\ResourceAwareGridInterface;
 
 class BookGrid extends AbstractGrid implements ResourceAwareGridInterface
 {
