+++ src/Component/Tests/Dummy/NoResourceGrid.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Sylius\Component\Grid\Tests\Dummy;
 
-use Sylius\Bundle\GridBundle\AbstractGrid;
 use Sylius\Bundle\GridBundle\Builder\GridBuilderInterface;
+use Sylius\Bundle\GridBundle\Grid\AbstractGrid;
 
 final class NoResourceGrid extends AbstractGrid
 {
